pyproject.toml
[project]
name = "kume"
version = "0.2.0"
description = "A multimodal AI agent for personal nutrition guidance via Telegram"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14",
    "asyncpg>=0.30",
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "langchain-postgres>=0.0.12",
    "pymupdf>=1.24",
    "python-dotenv>=1.0",
    "python-telegram-bot>=21.0",
    "sqlalchemy[asyncio]>=2.0",
]

[project.optional-dependencies]
dev = [
    "aiosqlite>=0.20",
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
    "mypy>=1.13",
    "pre-commit>=4.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "SIM108", "SIM105", "SIM102", "UP042"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
]
